Update "doGet(javax.servlet.http.HttpServletRequest,javax.servlet.http.HttpServletResponse)" method in "javax.servlet.http.MainServlet" class

+++ src/main/java/javax/servlet/http/MainServlet.java
@@ -146,8 +146,12 @@ public class MainServlet extends HttpServlet {
 				//
 			} finally {
 				//
-				FileUtils.delete(file);
-				//
+				if (file != null) {
+					//
+					FileUtils.delete(file);
+					//
+				} // if
+					//
 			} // try
 				//
 		} // if
